backend/requirements.txt
google-generativeai
fastapi
fastapi[all]
pydantic
python-dotenv
uvicorn
pydantic-settings
markdown